[BUGFIX] Test Fixtures
